Refactor terminal decoder to class
